Implement Challenge 38: Dictionary attack on SRP
Added client, server and attacker scripts for implementing "Challenge
38: Offline dictionary attack on simplified SRP".

The challenge is implemented through 3 scripts:

 * The server ('cp_test_c38_server.py'):
    * Implements the challenge's real server side protocol for the
      simplified SRP, as described by the challenge's statement, using a
      TCP server.
    * Must be started before running the client.
    * To be used for testing the simplified SRP protocol and confirming
      user/password validation.
    * Takes as arguments the host and port where it listens to requests
      (e.g. 'localhost' and '9001') together with the user/password to
      validate.
    * "Ctrl+C" can be used to abort the server.

 * The attacker ('cp_test_c38_attacker.py'):
    * To run instead of the server as an attacker impersonating the
      server and performing the attack prescribed by the challenge.
    * Implements a "hacked" version of the server side protocol. To the
      client, it appears like a normal valid server, so that it can
      control all the keys and parameters in order to capture a valid
      HMAC from the client's authentication request. It responds "OK" to
      the client without any validation (password is not know yet). It
      then executes a brute force, offline attack on the password (by
      testing all words from a dictionary) until an HMAC equal to the
      one given by the client is found, meaning the password has been
      unveiled.
    * Must be started before running the client and instead of the real
      server.
    * The arguments are the host and port where the attacker server
      listens to.
    * "Ctrl+C" can be used to stop the attacker server.

 * The client ('cp_test_c38.py'):
    * Implements the challenge's client side protocol.
    * Can be used to connect directly to the real server (e.g., for
      testing the protocol), but in the challenge's context, it connects
      to the attacker server instead.
    * The server that it connects to (either the real or attacker) must
      be running.
    * The arguments are the server (either real or attacker) host and p
      port plus user/password.

The client uses the 'socket' module to connect to the server configured
by the script's arguments (can either be the real server or the attacker
server). Then, it implements the protocol as described in the
challenge's statement and collects all relevant data built in the
meanwhile (the keys and parameters used, etc.). All script arguments are
displayed as well as the results and relevant data.

The real server uses the TCP Server utility classes to establish a
server listening on the host and port specified by the script's
arguments. Upon new connection requests, it implements the server side
protocol as described in the challenge's statement, validating the
client's user/password and providing the associated "OK"/"NOT OK"
result.

The attacker server is in all similar to the real server (the client
won't tell the difference) and implements the protocol correctly, with
the following differences:

 * Selects small, fixed values (instead of random) for 'salt', 'b' and
   'u'. This will speed up the calculations required for breaking the
   password, thus reducing the time required for breaking it.

 * Does not calculate 'v', 'S', 'K' and the server HMAC for validating
   the password, because it can't (password is not know while the
   protocol is being executed).

 * Doesn't actually validate the client's HMAC (again, it can't). It
   simply responds "OK" unconditionally.

 * Stores all relevant parameters from the protocol that will be
   required for breaking the password. These are 'N', 'g', 'salt', 'A',
   'b', 'u' and the client's 'HMAC'.

When the protocol with the client is completed, the attacker spawns a
new thread for performing the offline dictionary attack on the password,
provided with the required parameters saved during the protocol
execution. The thread is implemented using the 'PasswordBreaker' class,
which subclasses python's 'threading.Thread'. The constructor gets the
saved protocol parameters and then the 'run()' method (thread's main
function) executes the dictionary attack:

 * Uses '/usr/share/dict/words' as the dictionary for all the possible
   passwords to be tested.

 * For each word, uses the protocol saved parameters for calculating
   'v', 'S', 'K' and the server's HMAC. The password is unveiled when
   the calculated server's HMAC is equal to the saved client's.

It should be noted that the attacker will fail to break the password if
it's not a word contained in the used dictionary.

Example usage for the scripts:

 * Use only the client and real server:
    `python cp_test_c38_server.py localhost 9001 rjpdasilva butter`
    `python cp_test_c38.py localhost 9001 rjpdasilva butter`

 * Use client and attacker server:
    `python cp_test_c38_attacker.py localhost 9001`
    `python cp_test_c34.py localhost 9001 rjpdasilva butter`

All scripts provide some output on what's going on and the client
script provides, as usual, info about relevant data and overall test
result. To note that, when used with the attacker, the client will
always see "OK" for thee authentication result. In this case, the test
success is actually in the attacker, which will show the results of
password breaking.
There's a 'debug' flag (defaults to 0) on all scripts' code
which can be set to '1' to enable more verbosity and debug messages.

The 'README.md' file has also been updated with new data brought in by
this challenge (new usage, dependencies and example).