Implement Challenge 42: Bleichenbacher RSA Attack
Added script for implementing "Challenge 42: Bleichenbacher's e=3 RSA
Attack".

The attack, as described by the challenge's statement, is performed in
several steps, controlled by the 'execute_forge_signature()' function
called from the main script which gives it the message to use and the
RSA keys bit size:

 * Generate the required RSA key pair with 1024 bit keys, done by
   'gen_keys()'.

 * Perform the real message signing using 'sign_real()' with the message
   to sign, the RSA private key and respective bit size (required for
   padding). Also checks if the signature is valid by calling
   'sign_verify()' with the RSA public key, the message and the real
   signature to check.

 * Do the RSA signature forging using 'sign_forge()', which only gets
   the message for signature forging and the RSA keys bit size (required
   for padding). It also checks if the forged signature is valid using
   the 'sign_verify()' function. The implementation follows the
   challenge's statement recommendations and comments/references were
   added for explaining the attack's rationale.

 * Gather the information required to be returned to the main script,
   which is the real and forged signatures together with their
   respective signature check results.

The main script gets the test data from the 'execute_forge_signature()'
function, presents it to the user and asserts the test success by
checking the both the real and forged signatures we're both successfully
verified. For testing and debugging, additional verbosity can be enabled
by setting the global 'debug' flag to 1.

The core modules dependency list on 'README.md' has been updated to
include the 're' (regex) module used by this challenge.