Implement Challenge 45: DSA parameter tampering
Added script for implementing "Challenge 45: DSA parameter tampering".

Main script calls 'execute_test_tampered_g()' with the test parameters
'(p, q, g)' and with the two test messages. The function is called once
with 'g = 0' and a 2nd time with 'g = p + 1' as required by the
challenge. It simply returns success or failure to the main script,
which presents the results and asserts the test success.

The 'execute_test_tampered_g()' function does the following:

 * Generate a DSA key pair and confirm that 'y' from the public key will
   have a forced value of '0' when 'g = 0' or '1' if 'g = p + 1'.

 * Sign both messages with the generated keys and confirm that the 'r'
   values from the signatures have, just like 'y', the same forced value
   depending how 'g' is being tampered.

 * Confirm the signatures are verified correctly, using the correct
   signatures previously generated. This is a normal case.

 * Confirm that the signatures also verify, even when switching the
   signatures for the messages, i.e., using 'sig2' for 'msg1' and 'sig1'
   for 'msg2'. This is an abnormal case explained by the 'g' value
   tampering.

 * Confirm that any '(r, s)' signature is also accepted for any value
   of 's' as long as 'r' is fixed to the expected forced value ('0' when
   'g = 0' or '1' if 'g = p + 1'). This is also an abnormal case due to
   'g' parameter tampering. Due to time limitation, "only" 10000 random
   values for 's' are tried.

The attack rationale is explained in the 'execute_test_tampered_g()'
function initial comments.

To note that the standard DSA sign and verify functions will reject
signing/verifying when 'r = 0' and/or 's = 0'. The initial
implementation of the utility functions 'dsa_sign()' and 'dsa_verify()'
did it as well. When running this challenge with 'g = 0', the initial
implementation would reject signing/verifying messages, because 'r' will
always be '0'. This way, their implementation was extended with a bool
'strict' parameter (defaulting to 'True') to allow bypassing those tests
for this challenge's purpose, i.e., the challenge will set the new
argument 'strict = False' when signing and verifying with 'g = 0'.

The 'c43.hash_msg_as_int()' was reused from Challenge 43, used for
calculating the message hashes as integers.