Implement Challenge 46: RSA parity oracle
Added script for implementing "Challenge 46: RSA parity oracle".

The attack is performed as described in the challenge's statement, split
into two major steps.

The 1st step is the data capturing, performed by the function
'execute_capture_data()', called from the main script with the base64
encoded plaintext and the RSA keys bit size. The function generates a
new RSA key pair, encrypts the plaintext message and returns the RSA
public key plus the ciphertext back to the main script, as the captured
data.

The main script takes the captured data and gives it to the
'execute_break_rsa()' function for the 2nd step, where the actual
ciphertext breaking is performed following the process described by the
challenge's statement. A detailed explanation is given for the attack in
the function's comments, explaining the rationale behind it as well as
its implementation. As prescribed by the challenge, the current status
of the broken string is shown as the breaking process goes on, starting
with garbage, but then getting nearer and nearer the final message as
the process develops till its end.

The 'execute_break_rsa()' function returns the broken message back to
the main script, which presents the final results and accesses the test
success, by comparing the returned broken message against the original
version of the plaintext.

Additional verbosity can be enabled by setting the global 'debug' flag
to 1.