Implement Challenge 47: B98 Pad Oracle (simple)
Added script for implementing "Challenge 47: Bleichenbacher's PKCS 1.5
Padding Oracle (Simple Case)".

The attack is performed as described in the challenge's statement,
following the steps and procedures in the Bleichenbacher's paper.

The challenge, executed from the 'main()' function, called from the main
script, is split into 2 major phases:

 * Data Capturing:
    Done by the 'execute_capture_data()' function. Receives the message
    to encrypt and the RSA keys bitsize, generates a new RSA key pair,
    pads the message (PKCS#1 v1.5) and encrypts it using the private
    key. The encrypted message and the RSA public key are returned by
    this function, to be used as input for the attack.

 * Message Breaking:
    Done by 'execute_break_rsa()' and implementing the attack as
    described by the challenge and the Bleichenbacher's paper.
    Receives as input the data captured (message ciphertext and RSA
    public key) and produces as output the plaintext broken message (if
    any) and some statistical information (time took to break the
    message and number of iterations).

The message breaking function takes several steps, matching the steps
described in the Bleichenbacher's paper

The first step is "Step 1: Blinding" and it's used for the
initialization of the process variables ('k', 'B', 's0', 'M', etc.).

The remaining of the process is performed in a loop for narrowing down
the possible ranges for the message to break till it can only have one
value, i.e., the loop ends when there's only one possible interval for
the message and that interval is such that it contains only one value.

The loop includes:

 * Step 2: Searching for PKCS conforming messages:
    The process of finding the next 'si' value, based on the current
    state, that is able to produce a confirming PKCS message.
    It is in this step that the 'decryption_oracle()' oracle is used for
    determining if a new message is PKCS conforming.
    Done by the 'find_si_next()' function, which itself can follow 3
    ways, depending on the current state:

    * Step 2a: Starting the search (i = 1):
       Used only in the 1st loop iteration. Uses the function
       'find_si_normal()' with an initial 's' of 'n/3B' to find the next
       'si' (trying one by one, incrementing) that produces a conforming
       PKCS#1 v1.5 message.

    * Step 2b: Searching with more than one interval left:
       Used when there's currently more than one possible interval for
       the message to break. Also uses 'find_si_normal()' for finding
       the next 'si', but uses as initial 's', the previous 's' plus
       one.

    * Step 2c: Searching with one interval left:
       Done when there's currently more than one possible range for the
       message to break. This is the case where we can speed up the
       search (i.e., narrow the intervals faster by getting larger next
       'si' faster). The faster method of getting the next 'si' is done
       in the 'find_si_quick()' function.

 * Step 3: Narrowing the set of solutions:
    Recalculate the possible set of intervals for the message to break
    based on the newly calculated 'si' from step 2. Implemented in the
    'find_mi_next()' function.

When the loop completes, the final "Step 4: Computing the solution" is
entered, which just determines the broken plaintext and performs some
sanity checks on the broken message padding validity.

Besides the functions already mentioned, the script contains the
following additional ones:

 * 'int_from_bytes()' and 'bytes_from_int()': For converting byte arrays
   to integers and vice-versa.

 * 'pkcs1_15_pad()' and 'pkcs1_15_unpad()': Perform the PKCS#1 v1.5
   padding and unpadding of messages, including correct padding
   validations. The unpadding function is also used by the decryption
   oracle to get the information it is required to leak to the attacker
   (if a decrypted ciphertext starts by '0x00 0x02' or not).

 * 'dump_intervals()': Debug function to show current possible intervals
   for the message being broken.

 * 'get_cipher_product()': To get a new ciphertext 'ci', based on a
    previous one 'c', using a multiplicative factor 'si', so that the
    associated plaintexts obey to 'mi = (c * si) % n'.

 * 'cipher_is_pkcs1_15_ok()': The main interface function used for
   knowing if a ciphertext is PKCS confirming. Just requests the info to
   the decryption oracle.

The 'main()' function starts by calling the data capture function. Then,
if all goes well, it provides it to the message breaking function.
Finally, it presents the results and determines the test success by
confirming that the broken message matches the original plaintext.
Additional verbosity can be enabled by setting the global 'debug' flag
to 1.

The main script just prepares the arguments (message and keys bit size)
for the 'main()' function and calls it.

The reason for having this very thin main script and a 'main()' function
is due to the next challenge, which exactly like this one, except that
the bit size will be different. This way, the next challenge can simply
import this one and call its 'main()' function the specific arguments.

In fact, according to this challenge's statement, the "Step 2b" would
not be required, because the key bit size of 256 doesn't normally cause
a scenario where that step is needed (although in some cases, it does).
However, the way this challenge was implemented already contemplates
that case, so it became a generic implementation working for all cases,
like the one on the next challenge, thus the choice of using a 'main()'
function that can be imported by the next challenge.