Implement Challenge 48: B98 Pad Oracle (complete)
Added script for implementing "Challenge 48: Bleichenbacher's PKCS 1.5
Padding Oracle (Complete Case)".

This implementation is fully reusing the script for Challenge 47, by
simply calling its 'main()' function with a modulus bit size of 768, as
required by the challenge.

The difference between C47 and C48, caused by the increased modulus bit
size, is that C48 needs "Step 2b" and needs "Step 3" to deal with
multiple intervals (steps from the BB'98 paper), whereas C47 would not,
in principle, need those features. In practice, for some values of 'n',
even with 256 bits, the steps above were also required in C47.

The way C47 was implemented was already for the generic attack process,
so it already included all the required features and steps. This way,
C47 and C48 can share everything, except the modulus bit size, so this
C48 implementation reuses everything from C47, giving the new bit size
of 768 as argument to the C47 'main()' function.