Implement Challenge 49: CBC-MAC Message Forgery
Added script for implementing "Challenge 49: CBC-MAC Message Forgery".

This challenge is split into two scenarios. In the 1st scenario the
attacker can control the CBC Initialization Vector (IV) and thus, able
to control as well the message's 1st block. The 2nd scenario forces the
IV to a fixed value (zero), but it's still possible to forge a message
by doing a length extension attack.

The main functions in this script are suffixed with the scenario number
"_X", either "_1" or "_2":

 * 'server_process_request_X()': Simulates the server receiving and
   processing a transfer request. Does some validations on the requests'
   syntax.

 * 'client_send_request_X()': Simulates the client issuing a transfer
   request to the server. Performs some validations on the request
   syntax. Also provides the final request message sent to the server as
   a means for the attacker to capture messages.

 * 'execute_forge_cbc_mac_X()': Perform the actual attacks prescribed by
   the challenge's statement. Contain a detailed explanation about each
   attack rationale. Provide the attack success result as well as the
   message request used for making the attack.

There's also an utility 'cbc_mac()' function for calculating the
CBC-MAC digests used throughout the challenge.

The main script calls both the 'execute' function in turn, collects
their outputs, presents input and output data to the user and determines
the test result by checking if both attack 'execute' functions returned
success.

Additional verbosity can be enabled by setting the global 'debug' flag
to 1.